Some fixes
- Move getIdFromUrl to utils
- Check for trailing slash in index url or add it
- Remove hardcoded values from getLink
- Give proper error message when file id or file metadata not found in getlink
