+++ src/.constants.js.example
@@ -6,6 +6,8 @@ module.exports = Object.freeze({
   ARIA_FILTERED_DOMAINS: ['yts', 'YTS', 'cruzing.xyz', 'eztv.ag', 'YIFY'], // Prevent downloading from URLs containing these substrings
   ARIA_FILTERED_FILENAMES: ['YIFY'], // Files/top level directories with these substrings in the filename won't be downloaded
   ARIA_PORT: 8210, // Port for aria2c RPC server, if you change this here, make sure to update aria.sh as well
+  API_ID: '',
+  API_HASH: '',
   GDRIVE_PARENT_DIR_ID: 'id_of_Drive_folder_to_upload_into',
   OTHER_GDRIVE_DIR_IDS: [], // This is needed if u want to look for files in multiple dirs on list command
   SUDO_USERS: [012, 345],	// Telegram user IDs. These users can use the bot in any chat.
@@ -27,5 +29,5 @@ module.exports = Object.freeze({
   },
   IS_TEAM_DRIVE: false,
   USE_SERVICE_ACCOUNT: false,
-  INDEX_DOMAIN: "" // add an extra / at the end, for example https://www.gdindex.url/
+  INDEX_DOMAIN: ""
 });
